Removing extra checks that are not needed. Removing unused ParentTypeSetter
